change in form request and migration to table 45 - solve bug fixed

+++ database/migrations/2022_06_22_205453_create_gostaresh_social_health_status_analyses_table.php
@@ -31,11 +31,11 @@ return new class extends Migration
             $table->string('unit')->nullable();
             $table->unsignedTinyInteger('component')->comment('مولفه')->nullable();
             $table->unsignedTinyInteger('gender_id')->unsigned()->nullable();
-            $table->unsignedSmallInteger('associate_degree')->nullable();
-            $table->unsignedSmallInteger('bachelor_degree')->nullable();
-            $table->unsignedSmallInteger('masters')->nullable();
-            $table->unsignedSmallInteger('professional_doctor')->nullable();
-            $table->unsignedSmallInteger('phd')->nullable();
+            $table->unsignedInteger('associate_degree')->nullable();
+            $table->unsignedInteger('bachelor_degree')->nullable();
+            $table->unsignedInteger('masters')->nullable();
+            $table->unsignedInteger('professional_doctor')->nullable();
+            $table->unsignedInteger('phd')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
